Fix "ready" promise to handle sync and async loading by Emscripten
If we compile without WASM, loading it's synchronous.
If we compile with WASM on, loading is async.

But we need a mechanism we can rely upon to tell us when the loading is complete, regardless of the underlying loading code.
So, we now have a "ready" promise that we can listen for on the exported module.

Also